Code refactoring, moving to common location Code refactoring, using correct assertion matcher Refactoring, renaming test file Minor code refactoring Code refactoring, adding reusability Update to spelling
