+++ app/core/utils/viewHelper.js
